Extract UserDataStoreProvider interface for class delegation
This commit introduces a `UserDataStoreProvider` interface to abstract the data store operations.

- A new `UserDataStoreProvider` interface is created, defining the contract for accessing user data like basic info, contest info, submissions, and the daily problem.
- `UserDatastore` now implements this new interface, and its public methods are marked with `override`.
- `UserDetailsRepositoryImpl` now utilizes delegation (`by userDatastore`) to implement the `UserDataStoreProvider` interface, removing boilerplate code that previously just called the corresponding `UserDatastore` methods.
- The method `getUserAcSubmissions()` has been renamed to `getUserRecentAcSubmissions()` in the interface and all implementing classes for better clarity.
